Use UPPERCASE filenames
On case sensative filesystems (e.g. linux) we have logic that if a file
fails to load, it re-tries with a lowercase filename (as some systems
seem to mount the apoc iso as lowercase, others uppercase). This means
the first attempted filename must be uppercase to try both.

Should fix loading the citymap if you're on such a system and your data
filenames are all uppercase