Load buildings file and allow debug tile selection
Now buildings are loaded and tiles are assigned accordingly. Added a
debug option to select a tile (Enable the debug checkbox in the main
menu, and control the selected tile with W/S A/D R/F in the mapview).

Information about the currently select tile is printed to console.