Increase data cache size
Due to the way ImageSets are (poorly) implemented, if the set is not in
the cache it is deleted, orphaning all sub-image. This causes it to hit
the 'non-batching' slowpath in the renderer.

Increase the cache size as a temporary work-around until I can get
around to fixing Image/ImageSet references properly