Add UString::c_str() returning a const char*
UString.str().c_str() is used enough to make this worth having.

This keeps a std::string locally to the UString to reference, as
the reutrned pointer needs to be valid until the UString is modified
(actually, until either the UString is called, or the next .str() call).

The fact that the result of str().c_str() was never used outside the
scope the .str() string was constructed in means that this just happened
to be safe, that that seems rather fragile.