Pathing: Allow origin==dest and always move at least one tile if routing fails
If every route is blocked and the 'closest tile' was the origin it would
loop forever as the goto mission instantly 'completes' and routes to
exactly the same tile again. Instead always move *somewhere* :P

And origin==dest happens if you try to goto the current building (or, in
the future, if you goto the current location) - currently the building
one takes off then isntantly lands, which is a bit useless but there you
go.