Extend 'supports' to adjacent tiles in x/y if nothing in z
This means that tiles that are not directly supported by anything
beneath are 'supported' by any adjacent scenery tiles in x or y.
If /any/ of the 'supporting' tiles are destroyed, this tile will
collapse.

This may not be accurate to vanilla, as adjacent tiles do not always
mean they are 'connected'.