Solve warning about collapsing tile with no object
This was caused by mutually-supporting tiles - if one was hit it'll
collapse what it supports, which would then collapse this (adding it to
the fallingScenery list).

The tileobject of the scenery tile actually hit would then be removed,
while this is still in the fallingScenery list, causing update (and the
warning) to be called.

Instead  remove the tileObject before collapsing supported scenery, so
this->collapse() becomes a NOP (as it checks for a nullptr tileObject to
mean 'already destroyed' (and therefore any collapsing already completed)