Split tile rendering out into 'Layers'
This changes the tileview to render a full Z level of each 'level' in
order, so (At the same Z) a higher 'layer' will always be on top of a
lower 'level'.

The scenery is drawn as the first layer, everything else (sorted by Z as
before) is then drawn on top of this. This fixes issues where shadows &
projectiles can be occluded by tiles in increased x/y when they overlap
a tile boundary.

May need some tweaks on how the 'cityovl' stuff works, otherwise some
things (e.g. vehicles taking off) may not may be drawn on top of things
they shouldn't (But this should be handled by the 'z' sorting, assuming
they have the correct position within a tile)