Don't draw shadows that have nothing to fall on
e.g. for landing pads this stops the shadow being drawn at the 'bottom'
of the hole (where the map ends) - as shadows are just drawn 'on top'
of any scenery with the same z, this causes these shadows to be drawn
over the top of scenery that should occlude it.

It's likely there are other cases where 'just draw shadows on top' is
wrong, but this one is pretty obvious