Add a relations map to Organisations, make isHostile compare to -50
Skin36 (Thanks!) found that the relations are organised thus:
75...100   allied
74...25    friendly
24...-24   neutral
-25...-49  unfriendly
-50...-100 hostile

clamped between (100,-100)
To track this we add a map of floats for each organisations relation with
other orgs.
There is a second map in the original code - likely the 'initial' relation
(which I think it will slowly tend to with no other input)

As this relationship map is currently empty, it will be initialised to the
default (0) at first use, so effectively everone is neutral. This means that
they will no longer shoot everyone in the cityscape.