Fix CITYMAP_ALIEN building names, fixes vehicles 'getting stuck'
albuild.bld seems to use a different naming index scheme to citymapN.bld?

This was causing id collisions with buildings in the human/alien city maps, where it was trying to 'land' at the x/y position of a building that doesn't exist in the current city.

TODO: This currently assumes the alien buildings are listed in name-order, need to check if that's correct