Add --load-save and --frame-limit options
Useful for benchmarks, or instantly loading a save (can use as handler
for openapoc save files? :)

Syntax: --load-save=save - tries to load 'save' and 'save.zip'. You'll
probably have to quote the whole arg (e.g. '--load-save=my save/path' to
get it in the same arg. Also splits on '=', so paths with an '=' will
break...

Similarly, --frame-limit=5 - if it failed to parse the int it'll behave
like the default '0' was specified (0 = no limit)