Fix GotoBuilding mission that was stopping things landing
As isFinished() would return true when the 'last' target has been popped
from the queue, this was causing the 'try to land' logic being run when
the vehicle starts moving to the last tile, not when it reaches the last
tile.

This means it would rarely actually land, as it would be unlikely a
valid landing pad would be under the second-from-last tile in the path.

Now this only sets isFinished() when you try to get a new location after
the last