Just warn if a PCKSTRAT file has pixels outsize the 8x8 box
Some images (e.g. some sprites in the 48GATE map) appear to have
'negative' x (IE less than 640 stride) for some pixels. Not sure if this
is intentional, some small error in the source data, or something we're
still doing slightly wrong in our loader. But either way, don't error
out when we hit that.