gamestate_serialize: Enable comparison against ref value at serializeOut
This means if the value hasn't changed from the default we don't emit
the node.

Makes the resulting xml a bit cleaner.

TODO: Allow comparing reference map values with the same key. Similarly,
lists and vectors and the like. This may require some more complexity
(e.g. like detecting is a node is removed from the reference and
'delete' it)