Add StateObject::generateObjectID() to create unique object IDs
This stores an incrementing uint64_t in the GameState, which is
serialized, guaranteeing each generated ID to be unique for the lifetime
of the GameState.

Unless you start overflowing a uint64_t, which sounds pretty crazy to me
right now...

This fixes starting vehicle battles in debug mode, due to a failing
object ID prefix test.