Build fix for glm-0.9.8
This seems to have added Vec2<>->Vec2<> conversions, so now we need to
be a bit more careful with implicit conversions to make sure it chooses
the right one.

E.g. supplying {int, int} to something that expects a Vec2<float> would
implicitly convert the ints to floats, but now the compiler can't figure
that out