Add a 'version' string to the serialized GameState
Current it's not used for anything, but useful to eyeball to make sure
you're running against a 'correct' version. It's generated from a sha1
hash of the gamestate_serialize.xml, so standard sha1 tools can also be
used to verify.

In the future it may be used to block trying to load incompatible
formats.
If we add a 'extractor version' and some kind of source file hash we
could also use it to see if we actually need to re-run the extractor on
build.