SDLRaw audio: Expand sample rate conversion output buffer if necessary
Use the sdl-provided len_mult to see what size output buffer is
necessary.

This could waste memory if downsampling, as it'll keep the input buffer
size (but that's likely rare, we're normally upsampling 11/22khz sounds
to 44/48)

It also might be slightly slower, as it could cause a
init(input_size); resize(output_size);
instead of
init(max(input_size,output_size);
but it's better being correct than fast for now.