Framework: Fix ConfigFile static constructor use
Depending on what order the constructors are run at startup, relying on
the up<> constructor can break some settings, as the ConfigOption
objects are created, adding them to the ConfigFile instance, then the
up<ConfigFile> constructor is run, re-initialisig the up<> and losing
any config options that may have already be setup.

Using a 'raw' pointer instead of a up<> means it's instantiated in the
bss, so guaranteed not to be reset in the middle of the init
constructors.