MSVC: Default to 'OpenApoc' as the startup project
A number of people just load the solution and click 'run' - which in
msvc tries to run the "startup project", which (unless set) defaults to
'glm', which has no targets to build and gets confused trying to run a
non-existant static library. This option seems to be set in the 'user'
config for msvc, so not something we actually want in git anyway.

Googling suggests the first project is the start project, but then that
should be physfs not glm?