Make the data extraction an optional step in cmake
If you set the cmake var EXTRACT_DATA to off (default on), it no longer causes
the data extractor to be run as part of the default build. This means CD_PATH
would no longer need to point to a valid file, for example.

The extraction step can still be run with the 'extract-data' target - make sure
this is run before trying to run the game or any tests!