Use the "simple" SDL Dialog for errors
As that can apparently be called from any thread, use that instead - as
on some OSs (MacOS only?) this may cause crashes when called from a
thread other than that which created the initial window.

This has a side effect of no longer giving the "limp along" option - but
as the "error" level log is intended for fatal errors, it was nearly
always useless and gave users odd expectations.