Fix BattleMapPart destruction explosion doodad position
This seemed to be intended to be offset by x=-0.5 for the left wall from the
tile center, and -0.5y for the right wall (as the origin 0,0 is in the top-left)
but was double-negatived to invert that. I think that's wrong, and causes a
number of "out of bounds" doodads to be created, showing up as a number of
"Invalid Tile Coordinates" errors.

Hopefully this fixes that? A quick glance looks OK, but the position of the
doodads never quite matched up to my eye, even in the OG.