Workaround crashes on exit due to use-after-free of renderer
Some objects (like the ControlGenerator) keep references to Images in global
static scope, so end up being destructed at process termination time - after
the renderer has already been destroyed.

This then causes a use-after-free trying to callback into the renderer to
cleanup it's internal data.

Workaround this by keeping a static flag if that renderer is still alive.

I don't like this 'fix', as it'll break if there's ever more than 1 renderer
of each type constructed in the process lifetime, and things shouldn't keep
static handles to Images anyway, but this avoids the issue in the short-term