Don't create tileObject of vehicles between dimensions at game load
Vehicles currently in a dimension gate are currently assigned a tileObject in
GameState::initGame(), despite them not having valid position coords, so appear
on the map at a nonsense location (0,0,0) and hit a LogError() in
Vehicle::leaveDimensionGate() as it doesn't expect an existing tileObject.

So instead store a flag for if a Vehicle is between dimensions - and don't
create TileObjects for Vehicles with that flag set at game load time.

Note: This doesn't fix existing saves where there are vehicles currently in the
gates, as the flag won't be set correctly. New games, or games loaded from a
state where there aren't any Vehicles that happen to be in the gates at that
moment, should be fine going forward.