Snooze for a tick when restarting a GoTo mission that's blocked
I think this may be a possible cause of a deadlock - if a unit can't route any
closer to it's intended destination, it'll loop in Restart->Goto->Restart
forever.

Adding a 1 tick snooze should allow the game to progress, and possibly allow the
blockage to clear itself if it's temporary (e.g. another unit)