Lock the framework update loop to a set FPS (default 60)
This sleep()s if we're going faster than the expected update rate,
hopefully avoiding any possible issues due to different vsync limits
causing significantly different in-game speeds - as it's set to
progress 1 tick every frame.

Note that if you are 60fps vsync limited (or lower) there's a good
chance the update loop will 'fall behind' due to natural jitter and
precision differences, it'll LogWarning() if that happens, but it
shouldn't cause any real issues, so long as the difference is small.

If you somehow ended up with some weird setup with a significantly lower
than 60fps vsync, however, it will run slower.