WIP MacOS App bundle creation
Very much a WIP:
- Only the OpenApoc executable is packaged - we probably want /at least/
the launcher in there too
- The post-build copy cmake command for the data/ directory into the
resources feels like a hack - same with the fixed location .app bundle
name where it's copying them to? Probably some cmake property we can
query?
- It seems the working directory of app bundles is always "/"? Which
breaks all the loading "local" data, so we chdir() based on the SDL
"base directory", which seems to point to the Resources dir in the
bundle?
- Keying the above off the working directory being "/" is a hack - e.g.
run from the terminal it may be unlikely and maybe a bad idea, but
possible that "/" was the intended working directory - at which point I
don't really know that the SDL "base dir" would be?
- A whole stack of the app bundle metadata isn't set (authors,
"friendly" name, copyright etc.)