Avoid drawing vehicle equipment at (0, 0)
For the first render or so the position of the inventory screen
returns as (0, 0). This causes inventory items to be drawn at the
top-left hand corner of the screen. This makes the screen look
glitchy, so a simple check has been added to avoid drawing the
inventory items at all when the inventory screen reports its location
as (0, 0). Ideally, this will result in any strange draw calls being
ignored and the correct coordinates being loaded in before the
user notices anything is amiss.