Implement op="delete" attribute for maps and their entries
Map nodes with the op="delete" attribute are cleared before any
subsequent <entry> nodes are processed.

Map entry nodes with op="delete" remove the object with a matching key.
If a <value> node is supplied, a new default-constructed object is
created in that key slot, then the <value> processed as normal.