Check adjacent slots if dragging equipment to an invalid slot
This makes it way less frustrating to equip solders - less guessing
which x/y position that slot happened to be defined at (which is
visually off for the armor, certainly).

It current tries in the order (+x, -x, +y, -x, +x+y, +x-y, -x+y, -y-y),
and seems to do the 'right' thing most of the time.