HACK - Define UniChar as uint32_t
This works around things (E.g. the fmtlib api) that expect UniChar to be
an encoded character, while we always just treat it as an integer
codepoint

Hopefully this can be removed completely (along with much of the UString
class itself) when c++20 char8_t and std::basic_string<char8_t> actually
define a 'utf8 encoded string'.