Remove UString and just use std::string and std::u32string
We don't need to wrap the whole string, in c++20 we can use u8string
(std::basic_string<char8_t>) to ensure that it's a compile-time error to
assume that char (aka a 'byte') and char8_t (a utf8 code unit) are the
same.

Pull the few helper functions we implement on top of std::string into
loose functions.

If you want to use unicode codepoints (e.g. the font rendering or
textedit selection), use U32String instead.