Fix resizing converted audio buffer when changing frequency
This resize is wrong, as it doesn't take into account frequency changes.

Fixes issues with music when converting sample rate (e.g. higher than
the output 22050hz would cause it to be 'choppy' as this resize inserted
blank spaces in the buffer), or cut off sound effects if converting up.