Add getLanguage() and getLanguageCountry() to Framework
getLanguage should return the ISO-639 language code (e.g. "en" for
english, "fr" for french etc.)

getLanguageCountry should return the same as above, but (if set) appends
the ISO-3199 country code (e.g. "en_GB" for British English). Note that
that appending is optional - callers still have to handle "en" in that
case.

The goal is that when loading translated resources, we first try the
getLanguageCountry() version, fall back to getLanguage(), and then
assume it's not translated to that language.