cmake: Add USE_SYSTEM_QT option to use an installed QT package instead of vcpkg
As the vcpkg builds qt5 from source, this may be undesirable for some people - or the CI systems which have a max timeout that building qt5 from scratch blows past.

If set, the user needs to setup CMAKE_PREFIX_PATH to the installed qt5 directory/toolchain path manually if trying to build with qt.