Add support to the launcher to select languages (#1039)
* Add support to the launcher to select languages

ModInfos can now list a set of supported languages.
Currently, only the first enabled mod is used to populate this list.

The default is currently american english. I guess we could try to see
if the system locale is supported and use that as a default? Or is
english the expected default for games?