Fix build issue on new MSVC compilers
At least MSVC 19.29.30133 classes defined local to functions no longer see enum class types defined in the same parent function.

Not sure if this is an MSVC bug, or we were running fast and loose with the c++ spec other compilers aren't as strict about, but a small bit of file-local namespace pollution is probably fine.