Call Iterator destructors for objects created for lua (#1069)
The lua code provides iterators to recurse through maps, which directly
correspond to the c++ iterator objects. These are allocated with
lua_newuserdata, and a placement new called to initialise.

However, the destructors were never called. This caused issues on MSVC
debug builds, which have some container/iterator debugging that seems to
require the destructors to be called for some reference cleanup,
otherwise it crashes when destroying the parent container, as it tries
to access now-free'd memory.

This fix should add a __gc metatable method to the iterator objects
created in lua, which call the destructor beforee freeing, allowing
whatever cleanup to happen and fix the crash.

Fixes #1068