Load init lua scripts using Filesystem
The mod ScriptsList (scripts ran at mod init) didn't use the virtual FS, so would not load if the $(CWD) didn't match the data directory.

This caused warning as scripts failed to load in the tests, as well as if any mod overrides the file.

Note that this means that scripts loaded through dofile() in lua are now
relative to the data/ directory and can be replaced by mods overwriting
the name in their own data.