Multifix
Currently there is timing issue with the way events are processed. In
the case of the transferscreen ui elements are removed form a list and
added to another. In order to do this the UI library will remove a
shared pointer from a std::vector of controls. Doing this in a call back
creates a situation in which a control iterator becomes invalid. If the event
processing is in the process of iterating these controls this will
cause a crash.

Part of this commit introduces a removed flag to the control class. Now a list
box in the callback will instead mark the control with a removed flag.
This flag is then checked in the event iteration code. If true the event
loop will then be responsible for removing the element from the controls
list allowing the iterator to be updated at the same time preventing it
from becoming invalid.

This commit also fixes some other small issues in the transactionscreen
where foroms where not linked to correctly as well as adding guard
conditions to battleunit to account for the agent not always being set.