Update fmtlib, Replace sprintf-style "%" formats
A big change that replace all oldschool "%s" sprintf format with
fmt::format and it's "{}" style.

Breaks all current translations, as they encode the format strings
directly!

This is useful as it will allow positional arguments to be switched in
translations, though that is future work.

Updates fmtlib to 11.2.0, no particular reason just the latest tag at
the time of commit.

This rework was previously blocking fmtlib updates, as we used custom
type formatters with the 'legacy' sprintf fmt API, which is broken on
newer versions (and according to
https://github.com/fmtlib/fmt/issues/1499 only 'worked' in accident on
some veresions)

This was mostly auto-translated using clang-tidy and it's
'modernize-use-std-format' check - so may have issues. I've already
fixed up a few cases where float values don't get a max precision set
and overflow text controls.