Resurrect PCH support, enable on CI (#1529)
Update the PCH headers in each project and hook up to cmake.

Note this is disabled by default, as it seems to break incremental rebuilds when modifying files at times, but good for CI which build from scratch every time, and need all the help they can get WRT build time.

The contents of the PCH were a random assortment of the "most included files" of each project, minus some in GameState due to issues around instantiating templates before specialization (notable around StateObject and the ::get()/::prefix() templates)

The StateRef<T>::get() and similar templates might need reworking, including something that instantiates those before the definition is invalid, and that seems more likely to happen with PCHs and restricts the things we can precompile.